Replace LiskUser1234 address with SwiftyLisk address
Using two separate addresses, one for me (LiskUser1234) and one for SwiftyLisk makes for easier bookkeeping. At commit time I am the only maintainer of SwiftyLisk, but if in the future other people will join, it will be easier to send them part of the lisks donated to SwiftyLisk, because I will be able to distinguish between donations sent to me personally (LiskUser1234) and donations sent to the project  (SwiftyLisk).